refactor: 패키지 구조 변경 및 에러 응답 DTO 추가
