--- src/main/java/roomescape/common/exception/AlreadyInUseException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class AlreadyInUseException extends RuntimeException {
-
-    public AlreadyInUseException(final String message) {
-        super(message);
-    }
-}
