--- src/main/java/roomescape/common/exception/AuthenticationException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class AuthenticationException extends RuntimeException {
-
-    public AuthenticationException(final String message) {
-        super(message);
-    }
-}
