--- src/main/java/roomescape/common/exception/LoginFailException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class LoginFailException extends RuntimeException {
-
-    public LoginFailException(final String message) {
-        super(message);
-    }
-}
