+++ src/main/java/roomescape/common/exception/custom/PaymentBadRequestException.java
@@ -1,4 +1,4 @@
-package roomescape.common.exception;
+package roomescape.common.exception.custom;
 
 public class PaymentBadRequestException extends RuntimeException{
