--- src/test/java/roomescape/reservation/controller/WaitingControllerTest.java
@@ -1,160 +1,0 @@
-package roomescape.reservation.controller;
-
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.doNothing;
-
-import java.time.LocalDate;
-import java.util.Map;
-
-import io.restassured.parsing.Parser;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.annotation.DirtiesContext.ClassMode;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import roomescape.reservation.payment.dto.request.PaymentRequest;
-import roomescape.reservation.payment.service.PaymentService;
-
-@ActiveProfiles("test")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
-class WaitingControllerTest {
-
-    @MockitoBean
-    private PaymentService paymentService;
-
-    @DisplayName("예약 대기를 추가한다.")
-    @Test
-    void createReservationWaiting() {
-        int timeId = addReservationTime("10:00");
-        int themeId = addTheme();
-        String tokenValue = getAdminLoginTokenValue();
-        Map<String, Object> waitingParams = Map.of(
-                "date", getTomorrow(),
-                "timeId", timeId,
-                "themeId", themeId,
-                "paymentKey", "paymentKey",
-                "orderId", "orderId",
-                "amount", 1_000L
-        );
-
-        doNothing().when(paymentService)
-                .confirm(any(PaymentRequest.class));
-
-        RestAssured.given()
-                .cookie("token", tokenValue)
-                .contentType(ContentType.JSON)
-                .body(waitingParams)
-                .when().post("/reservations")
-                .then();
-
-        String userLoginTokenValue = getUserLoginTokenValue();
-        RestAssured.given().log().all()
-                .cookie("token", userLoginTokenValue)
-                .contentType(ContentType.JSON)
-                .body(waitingParams)
-                .when().post("/waitings")
-                .then().log().all()
-                .statusCode(201);
-    }
-
-    @DisplayName("예약대기를 삭제한다.")
-    @Test
-    void deleteReservationWaiting() {
-        int timeId = addReservationTime("10:00");
-        int themeId = addTheme();
-        String tokenValue = getAdminLoginTokenValue();
-        Map<String, Object> reservationParams = Map.of(
-                "date", getTomorrow(),
-                "timeId", timeId,
-                "themeId", themeId,
-                "paymentKey", "paymentKey",
-                "orderId", "orderId",
-                "amount", 1_000L
-        );
-
-        doNothing().when(paymentService)
-                .confirm(any(PaymentRequest.class));
-
-        RestAssured.given()
-                .cookie("token", tokenValue)
-                .contentType(ContentType.JSON)
-                .body(reservationParams)
-                .when().post("/reservations")
-                .then();
-
-        String userLoginTokenValue = getUserLoginTokenValue();
-        int waitingId = RestAssured.given()
-                .cookie("token", userLoginTokenValue)
-                .contentType(ContentType.JSON)
-                .body(reservationParams)
-                .when().post("/waitings")
-                .then().extract().path("id");
-
-        RestAssured.given().log().all()
-                .cookie("token", userLoginTokenValue)
-                .when().delete("/waitings/" + waitingId)
-                .then().log().all()
-                .statusCode(204);
-    }
-
-    @DisplayName("존재하지 않는 예약대기를 삭제할 경우 NOT_FOUND 반환한다.")
-    @Test
-    void deleteNonExistsReservationWaiting() {
-        RestAssured.given().log().all()
-                .when().delete("/waitings/0")
-                .then().log().all()
-                .statusCode(404);
-    }
-
-    private String getAdminLoginTokenValue() {
-        Map<String, String> adminLoginParams = Map.of("email", "admin@woowa.com", "password", "12341234");
-        return RestAssured.given()
-                .contentType(ContentType.JSON)
-                .body(adminLoginParams)
-                .when().post("/login")
-                .then()
-                .extract().cookie("token");
-    }
-
-    private String getUserLoginTokenValue() {
-        Map<String, String> adminLoginParams = Map.of("email", "user@woowa.com", "password", "12341234");
-        return RestAssured.given()
-                .contentType(ContentType.JSON)
-                .body(adminLoginParams)
-                .when().post("/login")
-                .then()
-                .extract().cookie("token");
-    }
-
-    private LocalDate getTomorrow() {
-        return LocalDate.now().plusDays(1L);
-    }
-
-    private int addReservationTime(final String timeValue) {
-        Map<String, String> timeParams = Map.of("startAt", timeValue);
-
-        return RestAssured.given()
-                .contentType(ContentType.JSON)
-                .body(timeParams)
-                .when().post("/times")
-                .then().extract().path("id");
-    }
-
-    private int addTheme() {
-        Map<String, String> themeParams = Map.of(
-                "name", "테마1", "description", "테마1", "thumbnail", "www.m.com"
-        );
-        return RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(themeParams)
-                .when().post("/themes")
-                .then().extract().path("id");
-    }
-}
